refactor: isolate optional LLM helper
